BUGFIX. Fix unittest, fix icon display in  project
